requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
cloudinary==1.41.0
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fonttools==4.54.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
imageio==2.36.0
jiter==0.7.0
kiwisolver==1.4.7
lazy_loader==0.4
matplotlib==3.9.2
networkx==3.4.2
numpy==2.1.2
openai==1.53.0
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
path==17.0.0
pillow==11.0.0
psycopg==3.2.3
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv
pytz==2024.2
response==0.5.0
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tifffile==2024.9.20
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
